requirements.txt
fvcore==0.1.5.post20221221

dgl
einops
ffmpeg
h5py
iopath==0.1.10
numpy==1.23.5
open3d
opencv-python==4.8.*
Pillow==10.0.0
pyheif==0.7.1
PyOpenGL
requests
scikit-image
scipy
seaborn
shapely==2.0.2
tqdm
trimesh==3.23.5
libigl
torchmetrics

ruff
pyright
pytest
pytest-cov

-f https://download.pytorch.org/whl/cu118
torch==2.0.1
torchvision

-f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py310_cu118_pyt201/download.html
pytorch3d
pydantic
smplx
rtree
torchmetrics
lpips
rembg